fix: Delete duplicate OCT 13, P&C fragment
P&C fragment doesn't seem to have broken link to "evidence tampering".
However, if SLED, or an outside forensic expert "Photoshopped" the digital image to highlight the blood splatter on the shirt, and the shirt was destroyed somehow during their analysis, then that evidence should not be considered. I hope that SLED retained the original digital image.